changed the type of attribute taille from int to double

+++ DashboardFitness/src/main/java/ma/emsi/dashboardfitness/entities/Utilisateur.java
@@ -20,7 +20,7 @@ public class Utilisateur {
 
     private String password;
     private int poids;
-    private int taille;
+    private double taille;
 
     @ManyToOne
     @JoinColumn(name = "idEntrainement")
